Replaced HeterogeneousAtmosphere with HeterogeneousAtmosphereNew
