+++ eradiate/scenes/atmosphere/__init__.py
@@ -1,6 +1,6 @@
 from ._core import Atmosphere, atmosphere_factory
+from ._heterogeneous_legacy import HeterogeneousAtmosphereLegacy
 from ._heterogeneous import HeterogeneousAtmosphere
-from ._heterogeneous_new import HeterogeneousNewAtmosphere
 from ._homogeneous import HomogeneousAtmosphere
 from ._molecules import MolecularAtmosphere
 from ._particle_dist import ParticleDistribution, particle_distribution_factory
@@ -8,8 +8,8 @@ from ._particles import ParticleLayer
 
 __all__ = [
     "Atmosphere",
+    "HeterogeneousAtmosphereLegacy",
     "HeterogeneousAtmosphere",
-    "HeterogeneousNewAtmosphere",
     "HomogeneousAtmosphere",
     "MolecularAtmosphere",
     "ParticleDistribution",
